Workaround fix for compile_erl function
Prior to this commit, cuttent ebin is added for erlc to run
with -pa. So the compiler can search current path when processing
-include_lib.
However, this makes erlang code server to cut off dir name
after the first hyphen to get app name.
e.g. if a erl app named foo_bar is in dir foo-bar, the app name
would be taken as foo regardless of what is in the .app file.

In this commit, path is added using -pz.
This is ok because it only affects include_lib
we should include local hrl files using -include
instead of -include_lib