Fix download relx failure
Using httpc to get release files from github results in 403 error:
`https://github.com/erlang/otp/pull/1859`

This causes downloading of relx failure when building emqx if the relx binary does not exist.

This PR fixes the validation for the existence of `wget`, and then uses `curl` as fallback solution.